pyproject.toml
[project]
authors = [
    {name = "Aleksey_Kuznetsov", email = "aleksejkuznecov07096@gmail.com"}
]
repository = "https://github.com/Mirroel-Alvares/python-project-83/tree/main"
name = "hexlet-code"
version = "0.1.0"
description = "Page analyzer"
readme = "README.md"
requires-python = ">=3.13"

classifiers = [
  'Development Status :: 3 - Alpha',
  'Environment :: Console',
  'Intended Audience :: End Users/Desktop',
  'Intended Audience :: Education',
  'License :: Free For Educational Use',
  'Natural Language :: English',
  'Operating System :: Microsoft',
  'Operating System :: Unix',
  'Programming Language :: Python :: 3.10',
  'Topic :: Software Development'
]

dependencies = [
    "bs4>=0.0.2",
    "flask>=3.1.0",
    "gunicorn>=23.0.0",
    "psycopg2-binary>=2.9.10",
    "python-dotenv>=1.0.1",
    "requests>=2.32.3",
    "ruff>=0.9.3",
    "validators>=0.34.0",
]

[tool.uv]
[dependency-groups]
dev-dependencies = []

#[project.scripts]
#gendiff = "gendiff.scripts.gendiff:main"
#
#[tool.hatch.build.targets.sdist]
#include = ["gendiff"]
#
#[tool.hatch.build.targets.wheel]
#include = ["gendiff"]
#
#[build-system]
#requires = ["hatchling"]
#build-backend = "hatchling.build"

